ingestion/requirements.txt
google-cloud-pubsub
websocket-client